+++ CHANGELOG.md
@@ -20,6 +20,7 @@
 - 完成实体的操作服务。
   - com.dwarfeng.familyhelper.life.impl.service.PbSetOperateServiceImpl。
   - com.dwarfeng.familyhelper.life.impl.service.PbNodeOperateServiceImpl。
+  - com.dwarfeng.familyhelper.life.impl.service.PbItemOperateServiceImpl。
 
 #### Bug修复